+++ src/extension.ts
@@ -5,6 +5,8 @@ import { installQuartoExtensionCommand } from "./commands/installQuartoExtension
 import { newQuartoReprexCommand } from "./commands/newQuartoReprex";
 import { ExtensionsInstalled } from "./ui/extensionsInstalled";
 import { getExtensionsDetails } from "./utils/extensionDetails";
+import { activateExtensions } from "./utils/activate";
+import { lintOnEvent } from "./commands/lint";
 
 /**
  * This method is called when the extension is activated.
@@ -38,6 +40,11 @@ export function activate(context: vscode.ExtensionContext) {
 	);
 
 	new ExtensionsInstalled(context);
+
+	if (vscode.window.activeTextEditor?.document.languageId === "quarto") {
+		activateExtensions(["DavidAnson.vscode-markdownlint"], context);
+		lintOnEvent();
+	}
 }
 
 /**